+++ src/Core/Shape.js
@@ -109,7 +109,7 @@ WHS.Shape = class extends WHS.Object {
 	 * @param {...String} tags - Tags that defines what to do with shape 
 	 * additionally.
 	 */
-	build( ...tags ) {
+	wrap( ...tags ) {
 		
 		'use strict';
 
@@ -506,7 +506,7 @@ WHS.Shape = class extends WHS.Object {
 
 		this.mesh = source.mesh.clone();
 
-		this.build();
+		this.wrap();
 
 		this.position = source.position.clone();
 		this.rotation = source.rotation.clone();