+++ src/Lights/HemisphereLight.js
@@ -33,8 +33,8 @@ WHS.HemisphereLight = class HemisphereLight extends WHS.Light {
                 params.helper.size ? params.helper.size : 0
             );
 
-        super.build();
-        super.buildShadow();
+        super.wrap();
+        super.wrapShadow();
 
 	}