+++ src/Shapes/Lathe.js
@@ -43,7 +43,7 @@ WHS.Lathe = class Lathe extends WHS.Shape {
             params.mass
         );
 
-        super.build();
+        super.wrap();
 
 	}